decompose fix
