+++ src/components/TodoList/TodoList.tsx
@@ -1,4 +1,4 @@
-import { Todo } from '../../../types/Todo';
+import { Todo } from '../../types/Todo';
 import { TodoInfo } from '../TodoInfo/TodoInfo';
 
 type Props = {
